refactor: move index.ts and config.ts into src
