+++ test/plugins/test-trace-datastore.ts
@@ -25,7 +25,7 @@ nock.disableNetConnect();
 describe('test-trace-datastore', function() {
   var agent;
   before(function() {
-    agent = require('../..').start({
+    agent = require('../../..').start({
       projectId: '0',
       samplingRate: 0,
       enhancedDatabaseReporting: true
