+++ test/plugins/test-trace-grpc.ts
@@ -279,7 +279,7 @@ Object.keys(versions).forEach(function(version) {
       });
 
       // It is necessary for the samplingRate to be 0 for the tests to succeed
-      agent = require('../..').start({
+      agent = require('../../..').start({
         projectId: '0',
         samplingRate: 0,
         enhancedDatabaseReporting: true,
