+++ test/plugins/test-trace-knex.ts
@@ -39,7 +39,7 @@ var versions = {
 describe('test-trace-knex', function() {
   var agent;
   before(function() {
-    agent = require('../..').start({
+    agent = require('../../..').start({
       projectId: '0',
       logLevel: 2,
       flushDelaySeconds: 1,
