+++ test/plugins/test-trace-koa.ts
@@ -35,7 +35,7 @@ describe('koa', function() {
   var agent;
 
   before(function() {
-    agent = require('../..').start({
+    agent = require('../../..').start({
       projectId: '0',
       ignoreUrls: ['/ignore'],
       samplingRate: 0
