+++ test/plugins/test-trace-mysql.ts
@@ -33,7 +33,7 @@ describe('test-trace-mysql', function() {
   var mysql;
   var pool;
   before(function() {
-    agent = require('../..').start({
+    agent = require('../../..').start({
       projectId: '0',
       enhancedDatabaseReporting: true,
       databaseResultReportingSize: RESULT_SIZE
