+++ test/plugins/test-trace-restify.ts
@@ -35,7 +35,7 @@ describe('restify', function() {
   var agent;
 
   before(function() {
-    agent = require('../..').start({
+    agent = require('../../..').start({
       projectId: '0',
       ignoreUrls: ['/ignore'],
       samplingRate: 0
