+++ test/test-config-plugins.ts
@@ -20,7 +20,7 @@ import * as pluginLoader from '../src/trace-plugin-loader';
 
 var assert = require('assert');
 var shimmer = require('shimmer');
-var trace = require('..');
+var trace = require('../..');
 
 var instrumentedModules = ['connect', 'express', 'generic-pool', 'grpc', 'hapi',
   'http', 'https', 'knex', 'koa', 'mongodb-core', 'mysql', 'pg', 'redis', 'restify'];
