+++ test/test-no-self-tracing.ts
@@ -35,7 +35,7 @@ describe('test-no-self-tracing', function() {
                 .get('/computeMetadata/v1/instance/hostname').reply(200)
                 .get('/computeMetadata/v1/instance/id').reply(200)
                 .get('/computeMetadata/v1/project/project-id').reply(200);
-    require('..').start({forceNewAgent_: true});
+    require('../..').start({forceNewAgent_: true});
     require('http'); // Must require http to force patching of the module
     var oldWarn = common.replaceWarnLogger(newWarn);
     setTimeout(function() {
@@ -51,7 +51,7 @@ describe('test-no-self-tracing', function() {
                 .get('/computeMetadata/v1/instance/id').reply(200);
     var apiScope = nock('https://cloudtrace.googleapis.com')
                 .patch('/v1/projects/0/traces').reply(200);
-    require('..').start({
+    require('../..').start({
       projectId: '0',
       bufferSize: 1,
       forceNewAgent_: true
