+++ test/test-plugins-no-project-num.ts
@@ -25,7 +25,7 @@ describe('test-plugins-no-project-num', function(){
   before(function() {
     savedProject = process.env.GCLOUD_PROJECT;
     delete process.env.GCLOUD_PROJECT;
-    agent = require('..').start();
+    agent = require('../..').start();
   });
 
   after(function() {
