+++ test/test-plugins-sample-warning.ts
@@ -31,7 +31,7 @@ describe('express + dbs', function() {
   var agent;
 
   before(function() {
-    agent = require('..').start({
+    agent = require('../..').start({
       projectId: '0',
       samplingRate: 0
     });
