+++ test/test-trace-options-sampling.ts
@@ -28,7 +28,7 @@ describe('express + mongo with trace options header + sampling', function() {
   var agent;
   var express;
   before(function() {
-    agent = require('..').start({
+    agent = require('../..').start({
       projectId: '0',
       samplingRate: 1
     });
