+++ test/test-unpatch.ts
@@ -21,7 +21,7 @@ import './override-gcp-metadata';
 var assert = require('assert');
 var nock = require('nock');
 var nocks = require('./nocks'/*.js*/);
-var trace = require('..');
+var trace = require('../..');
 
 nock.disableNetConnect();
