+++ tsconfig.json
@@ -8,7 +8,6 @@
     "noEmitOnError": true
   },
   "include": [
-    "*.ts",
     "src/*.ts",
     "src/plugins/plugin-https.ts"
   ]
